% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_css.R
\name{default_css}
\alias{default_css}
\title{Return the default CSS document used in the parallaxr package}
\usage{
default_css()
}
\description{
This returns a character vector of length 1 containing the default
CSS document used to generate the parallax scroll document.
}
