% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scroll_doc.R
\name{generate_scroll_doc}
\alias{generate_scroll_doc}
\title{Generate a HTML scroll document, passing a tibble with parameters as an argument}
\usage{
generate_scroll_doc(path, inputs, custom_css = NULL)
}
\arguments{
\item{path}{character string. Specify the path for the HTML file to be written.}

\item{inputs}{A tibble or data frame with three columns:
\itemize{
\item \code{content}: content in HTML to be written onto the page
\item \code{position}: either "left" or "right" to specify whether content
should appear on the left or right.
\item \code{imgurl}: URL for the image.
}}

\item{custom_css}{A character vector of length 1 that holds the CSS code for the parallax
document. You may use a combination of \code{readLines()} and \code{paste()} to create this vector
from a local CSS file in your working directory. Use \code{default_css()} to return the default
CSS to use for custom editing (you'll need to save the output into a custom CSS file).}
}
\description{
Pass a tibble with three columns to the \code{inputs} argument in
\code{generate_scroll_doc()}: \code{content}, \code{position}, and \code{imgurl}. The function then returns
a HTML document that will be saved in the location specified with \code{path}. The tibble
with parameters can be generated with \code{parse_md()} and \code{purrr::bind_rows()} (see
examples).
}
\examples{
\dontrun{
md_tibble <-
  purrr::bind_rows(parse_md("Markdown/1.md"),
                   parse_md("Markdown/2.md"),
                   parse_md("Markdown/3.md"))

generate_scroll_doc(path = "parallax-document.html",
                    inputs = md_tibble)
}

}
