% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2html.R
\name{md2html}
\alias{md2html}
\title{Convert Markdown text to HTML fragment}
\usage{
md2html(text)
}
\arguments{
\item{text}{Character vector containing Markdown text}
}
\description{
This is a wrapper around \code{markdown::markdownToHTML()}, where
the default behaviour is to produce a HTML fragment.
\code{htmltools::HTML()} is then used to evaluate the HTML code
within a RMarkdown chunk.
}
