% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_md.R
\name{parse_md}
\alias{parse_md}
\title{Parse a Markdown document with YAML headers and return a data frame for
generating a scroll document}
\usage{
parse_md(input_file)
}
\arguments{
\item{input_file}{character. File path to the Markdown document. The Markdown
document must contain a YAML header with the following keys:
\itemize{
\item position
\item imgurl
}}
}
\description{
Parse a Markdown document with YAML headers and return a data frame for
generating a scroll document
}
\examples{
\dontrun{
parse_md("Markdown/1.md")
}
}
