% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll_section.R
\name{scroll_section}
\alias{scroll_section}
\title{Generate the HTML div fragment for the scroll document}
\usage{
scroll_section(content = "right", position, imgurl)
}
\arguments{
\item{content}{character. Content in HTML}

\item{position}{character. Valid values include "right" (default) or "left".}

\item{imgurl}{character. URL string for the image to be displayed.}
}
\description{
This is designed to be used with \code{parse_md()} and
\code{purrr::pmap_chr()} to generate a HTML fragment for \code{generate_scroll_doc()}.
For the typical usage, you should not need to use this function directly.
This is exported for transparency reasons.
}
